Janky beginning of the parser
Going with a recursive descent for now. The AST nodes are currently not
the most beautiful of structures but I don't think we'll we able to use
the actual *Node for children because of terminals. Oh Well. Also, we
should really namespace out the tokens and node types into their own
packages so I don't need such awful names for everything. Whatever, such
is the price of forward progress.